chore: cleanup utils, cmd was the only package using it
